requirements.txt
# Core Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Logic Tensor Networks
logic-tensor-networks>=1.0.0
# Alternative LTN installation (use one of the above)
# ltn>=1.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0

# Bioinformatics
biopython>=1.79

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bars & Utilities
tqdm>=4.62.0

# Data Format Support
h5py>=3.6.0
pickle5>=0.0.11; python_version < "3.8"

# Optional: Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Advanced Visualization
plotly>=5.0.0

# Development Dependencies (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0 